// frontend/src/app/pages/private/administrative/app-admin-dashboard/shared/types.ts
export interface DashboardStats {
  totalUsers: number;
  totalUsersActive: number;
  totalNewUsers: number;
  totalUsersActiviedPercentage: number;
  totalUsersPercentage: number;
  totalNewUsersPercentage: number;
}

export interface ChurchStats {
  totalChurches: number;
  totalNewChurches: number;
  totalChurchPercentage: number;
  totalNewChurchPercentage: number;
}